fix: correct relative import paths in refactored components and hooks
Fixes remaining import resolution errors from incorrect relative paths:

TournamentDetailsSidebar.tsx:
- Changed: ../../../components/ActionButtonPair → ../../../../components/ActionButtonPair
- Changed: ../../../store/serviceApi → ../../../../store/serviceApi
- Reason: Component is nested one level deeper in TournamentId/components/

useTournamentEditHandler.ts:
- Changed: ../RegisterTournamentModal → ../AddTournamentModal
- Reason: AddTournamentModalRef is exported from AddTournamentModal, not RegisterTournamentModal

All import paths now correctly resolve from their source locations.
